feat: Fabric 1.19.4, split projects
