+++ forge/src/main/resources/pack.mcmeta
