+++ gradle.properties
@@ -1,2 +1,19 @@
-kotlin.code.style=official
-org.gradle.jvmargs=-Xmx5G
+# Done to increase the memory available to gradle.
+org.gradle.jvmargs=-Xmx1G
+org.gradle.parallel=true
+org.gradle.configureondemand=false
+
+# Fabric Properties
+# check these on https://fabricmc.net/develop
+minecraft_version=1.19.4
+yarn_mappings=1.19.4+build.2
+loader_version=0.14.19
+
+# Mod Properties
+mod_version = 1.0.0
+maven_group = com.example
+archives_base_name = fabric-example-mod
+fabric_kotlin_version=1.9.4+kotlin.1.8.21
+
+# Dependencies
+fabric_version=0.79.0+1.19.4
